+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Codeberg profile.
+- Home big preview swiper indicator.
+- Filter selected option indicator.
 
 ### Changed